Crate nds

Re-exports

pub use nds_sys as sys;

Modules

background

cache

debug

dma

API to control the DMA

embedded_graphics

input

interrupts

macros

sprite

system

video

Macros

dbg_breakpoint

Sets a NO$GBA style* breakpoint on debug builds. Before the breakpoint is hit, the filename and line number are print to the debugger console, along with an optional formatted message (akin to println!)

include_bytes_aligned

Includes the contents of a file as a statically allocated array with specified alignment.

include_bytes_as

Includes the contents of a file as a statically allocated array, casting it to an array of the specified type. While alignment is guaranteed (Check include_bytes_aligned!), the size of the array is rounded down to ensure it does not span past the end of the array.

print

println

Attribute Macros

entry

Use it to mark the ROM entry point.

panic_screen

Use it to mark a “screen of death” on run ROM.