Module nds::debug

Structs

NOCASH

Used to access all NO$GBA debugging functions. The debugger is detected on startup, and if it is found, NoCash::is_enabled will return true.

NoCash

Symbolizes the emulator NO$GBA and provides an API for its debugging capabilities. On release builds these functions don’t do anything; and when NoCash::is_enabled returns false, they return immediatly

Functions

_print

Writes a string byte-by-byte to the NO$GBA TTY. This function won’t allocate, so the caller has to take care to null-terminate their strings. Still, this function has 2 paths: If the string is null-terminated, it will be passed directly to NO$GBA, allowing for %param%s and faster code execution. Otherwise, is will be copied byte-by-byte to the Char Out

breakpoint

On debug builds this function sets a NO$GBA style* breakpoint. You should prefer the macro dbg_breakpoint!, since it provides debug info (and actually just calls this function)