Struct nds::debug::NoCash

pub struct NoCash { /* fields omitted */ }

Symbolizes the emulator NO$GBA and provides an API for its debugging capabilities. On release builds these functions don’t do anything; and when NoCash::is_enabled returns false, they return immediatly

Implementations

impl NoCash

pub fn new() -> Self

pub fn is_enabled(&self) -> bool

pub fn print_with_params_no_alloc(&mut self, s: &str)

Prints to the emulator TTY. NO$GBA will substitute %fmt% with a value according to this list (taken from gbatek):

r0,r1,r2,…,r15: show register content (displayed as 32bit Hex number)

sp,lr,pc: alias for r13,r14,r15

scanline: show current scanline number

frame: show total number of frames since coldboot

totalclks: show total number of clock cycles since coldboot

lastclks: show number of cycles since previous lastclks (or zeroclks)

zeroclks: resets the ‘lastclks’ counter

The caller must make sure the string is null-terminated

Doesn’t do anything unless NoCash::is_enabled() returns true

pub fn print_with_params<S: Debug>(&mut self, s: S)

Prints to the emulator TTY. NO$GBA will substitute %fmt% with a value according to this list (taken from gbatek):

r0,r1,r2,…,r15: show register content (displayed as 32bit Hex number)

sp,lr,pc: alias for r13,r14,r15

scanline: show current scanline number

frame: show total number of frames since coldboot

totalclks: show total number of clock cycles since coldboot

lastclks: show number of cycles since previous lastclks (or zeroclks)

zeroclks: resets the ‘lastclks’ counter

Doesn’t do anything unless NoCash::is_enabled() returns true

pub fn get_emu_id() -> Option<&'static str>

Returns a string identifying the emulator that the ROM is running on. Only NO$GBA supports this, so don’t rely on it