Function nds::dma::copy_words

pub unsafe fn copy_words(
    ch: Channel, 
    src: *const u32, 
    dst: *mut u32, 
    len: usize
)

Fills dst with len words of src. This function operates in words (32 bits), so the amount of bytes copied will be len*4

Safety

It is not recommended to use this function, it is only provided as a compat with the C version of libnds. This function works by instructing the DMA to copy len words from src to dst, so the caller must make sure that:

only [Copy] values are being copied,

[dst, dst + len) is valid and won’t overwrite unrelated data,

and the channel is free, since copy_words will overwrite it’s settings