Function nds::dma::fill_words

pub unsafe fn fill_words(ch: Channel, value: u32, dst: *mut u32, len: usize)

Fills dst with len copies of value. This function operates in words (32 bits), so the amount of bytes written will be len*4

Safety

It is not recommended to use this function, it is only provided as a compat with the C version of libnds. This function works by instructing the DMA to copy len times value words over dst, so the caller must make sure that:

only [Copy] values are being copied,

[dst, dst + len) is valid and won’t overwrite unrelated data,

and the channel is free, since fill_words will overwrite it’s settings