Struct nds::interrupts::Flags

pub struct Flags { /* fields omitted */ }

Implementations

impl Flags

pub const VBLANK: Flags

pub const HBLANK: Flags

pub const VCOUNT: Flags

pub const TIMER0: Flags

pub const TIMER1: Flags

pub const TIMER2: Flags

pub const TIMER3: Flags

pub const NETWORK: Flags

pub const DMA0: Flags

pub const DMA1: Flags

pub const DMA2: Flags

pub const DMA3: Flags

pub const KEYS: Flags

pub const CART: Flags

pub const IPCSYNC: Flags

pub const FIFOEMPTY: Flags

pub const FIFONOTEMPTY: Flags

pub const CARD: Flags

pub const CARDLINE: Flags

pub const GEOMETRYFIFO: Flags

pub const LID: Flags

pub const SPI: Flags

pub const WIFI: Flags

pub const ALL: Flags

pub const fn empty() -> Flags

Returns an empty set of flags.

pub const fn all() -> Flags

Returns the set containing all flags.

pub const fn bits(&self) -> u32

Returns the raw value of the flags currently stored.

pub const fn from_bits(bits: u32) -> Option<Flags>

Convert from underlying bit representation, unless that representation contains bits that do not correspond to a flag.

pub const fn from_bits_truncate(bits: u32) -> Flags

Convert from underlying bit representation, dropping any bits that do not correspond to flags.

pub const unsafe fn from_bits_unchecked(bits: u32) -> Flags

Convert from underlying bit representation, preserving all bits (even those not corresponding to a defined flag).

Safety

The caller of the bitflags! macro can chose to allow or disallow extra bits for their bitflags type.

The caller of from_bits_unchecked() has to ensure that all bits correspond to a defined flag or that extra bits are valid for this bitflags type.

pub const fn is_empty(&self) -> bool

Returns true if no flags are currently stored.

pub const fn is_all(&self) -> bool

Returns true if all flags are currently set.

pub const fn intersects(&self, other: Flags) -> bool

Returns true if there are flags common to both self and other.

pub const fn contains(&self, other: Flags) -> bool

Returns true if all of the flags in other are contained within self.

pub fn insert(&mut self, other: Flags)

Inserts the specified flags in-place.

pub fn remove(&mut self, other: Flags)

Removes the specified flags in-place.

pub fn toggle(&mut self, other: Flags)

Toggles the specified flags in-place.

pub fn set(&mut self, other: Flags, value: bool)

Inserts or removes the specified flags depending on the passed value.

pub const fn intersection(self, other: Flags) -> Flags

Returns the intersection between the flags in self and other.

Specifically, the returned set contains only the flags which are present in both self and other.

This is equivalent to using the & operator (e.g. ops::BitAnd), as in flags & other.

pub const fn union(self, other: Flags) -> Flags

Returns the union of between the flags in self and other.

Specifically, the returned set contains all flags which are present in either self or other, including any which are present in both (see Self::symmetric_difference if that is undesirable).

This is equivalent to using the | operator (e.g. ops::BitOr), as in flags | other.

pub const fn difference(self, other: Flags) -> Flags

Returns the difference between the flags in self and other.

Specifically, the returned set contains all flags present in self, except for the ones present in other.

It is also conceptually equivalent to the “bit-clear” operation: flags & !other (and this syntax is also supported).

This is equivalent to using the - operator (e.g. ops::Sub), as in flags - other.

pub const fn symmetric_difference(self, other: Flags) -> Flags

Returns the symmetric difference between the flags in self and other.

Specifically, the returned set contains the flags present which are present in self or other, but that are not present in both. Equivalently, it contains the flags present in exactly one of the sets self and other.

This is equivalent to using the ^ operator (e.g. ops::BitXor), as in flags ^ other.

pub const fn complement(self) -> Flags

Returns the complement of this set of flags.

Specifically, the returned set contains all the flags which are not set in self, but which are allowed for this type.

Alternatively, it can be thought of as the set difference between Self::all() and self (e.g. Self::all() - self)

This is equivalent to using the ! operator (e.g. ops::Not), as in !flags.