Macro nds::include_bytes_as

macro_rules! include_bytes_as {
    ($type:ty, $file:expr) => { ... };
}

Includes the contents of a file as a statically allocated array, casting it to an array of the specified type. While alignment is guaranteed (Check include_bytes_aligned!), the size of the array is rounded down to ensure it does not span past the end of the array.

Arguments

$type - The type to cast the included bytes as.

$file - The path to the file to include, see include_bytes!.

Examples

// Including the contents of "data.bin" as a slice of u32.
static DATA: &'static [u32] = include_bytes_as!(u32, "data.bin");

In the above example, if the file is 10 bytes long, the array will be 2 elements long, leaving the last 2 bytes unreacheable.

This macro is useful for including data that can be directly used by the hardware, such as palettes or bitmaps, or copied using DMA.