Enum nds::video::banks::C::Mode

#[repr(u8)]
pub enum Mode {
    Lcd,
    MainBgSlot0,
    MainBgSlot1,
    MainBgSlot2,
    MainBgSlot3,
    Arm7Slot0,
    Arm7Slot1,
    SubBgSlot0,
    TextureSlot0,
    TextureSlot1,
    TextureSlot2,
    TextureSlot3,
}

Type of mapping that can be used with Bank C.

Variants

Lcd

Maps Bank to LCD.

MainBgSlot0

Maps Bank to Main’s background slot 0. (Address: 0x06000000)

MainBgSlot1

Maps Bank to Main’s background slot 1. (Address: 0x06020000)

MainBgSlot2

Maps Bank to Main’s background slot 2. (Address: 0x06040000)

MainBgSlot3

Maps Bank to Main’s background slot 3. (Address: 0x06060000)

Arm7Slot0

Maps Bank to workram slot 0 of the ARM7. (Address: 0x06000000)

Arm7Slot1

Maps Bank to workram slot 1 of the ARM7. (Address: 0x06020000)

SubBgSlot0

Maps Bank to Sub’s background slot 0. (Address: 0x06200000)

TextureSlot0

Maps Bank to 3d texture slot 0.

TextureSlot1

Maps Bank to 3d texture slot 1.

TextureSlot2

Maps Bank to 3d texture slot 2.

TextureSlot3

Maps Bank to 3d texture slot 3.

Implementations

impl Mode

pub fn set(self)

Sets the specified mode on this bank while also enabling it