Module nds::dma

API to control the DMA

The Nintendo DS contains a DMA (Direct Memory Access) hardware that allows the console to copy and fill memory sections without using the CPU. The DMA can only respond to ONE channel at a time, and it will prioratize lower channels first. In case an operation is being processed, and a new one with a higher priority is received, the one with the lowest priority will be put on hold and the other will be fulfilled. 4 cycles must be waited after issuing a request before cancelling/overwriting it; (See the first 4 lines of wait_for); failing to do so may lead to a lock up (See point 1 below)

All functions here are NOT synchronous (except for wait_for), they return immediatly. Use is_busy and wait_for to check and wait for a channel to be available

Warning

The DMA can violate any safety guaranty in Rust, since it can be used to overwrite any memory segment using any value. It can be challenging to properly and safely use this hardware, even with this API. When developing, one should keep in mind the following things:

There is a delay of 2 cycles after issuing a request (Writing Flags::ENABLED) and the DMA actually starting. Don’t touch the channel during that period, it will lock up.

Toggling OFF the bit ENABLED will halt the DMA immediatly. This is not recommended unless the channel was programmed to autorepeat. In any case, wait at least 4 cycles after starting before halting the channel.

Wait until the channel has finished before issuing a command, otherwise the current operation will be overwritten (in the best case, see points 1 and 2 for the worst case). (copy and fill are the only non-unsafe functions here, and automatically call wait_for)

The hardware itself doesn’t have access to the CPU cache, therefore, data in the stack (such as local variables) may not be available yet in main memory. Static variables and Boxed values are ok, flushing the cache is also an option. (copy() and fill() take care of this issue by themselves)

Enums

Channel

The DMA has 4 “channels” to operate. Channels set the priority of operation being processed, so when multiple operations are scheduled, the one with the highest priority (= lowest channel) is resolved first. If another operation with a lower priority is being processed, then it is stopped and later resumed

Functions

copy

Copies src into dst using DMA channel 3. Calls wait_for on Channel::Ch3(Channel 3), so if interrupt DMA3 is disabled a panic may occur. Panics if size_of::<T>() is neither 2 nor 4. In case src.len() != dst.len() then only min(src.len(), dst.len()) elements will be copied.

copy_half_words⚠

Fills dst with len half words of src. This function operates in half words (16 bits), so the amount of bytes copied will be len*2

copy_within_unchecked⚠

Copies len elements from src. Starts copying at from, and copies to to. Hangs if the channel is busy. Panics if size_of::<T>() is neither 2 nor 4; but doesn’t do any bounds check

copy_words⚠

Fills dst with len words of src. This function operates in words (32 bits), so the amount of bytes copied will be len*4

fill

Fills dst by copying value using DMA channel 3. Copies src into dst using DMA channel 3. Calls wait_for on Channel::Ch3(Channel 3), so if interrupt DMA3 is disabled a panic may occur. Panics if size_of::<T>() is neither 2 nor 4.

fill_half_words⚠

Fills dst with len copies of value. This function operates in half words (16 bits), so the amount of bytes written will be len*2

fill_words⚠

Fills dst with len copies of value. This function operates in words (32 bits), so the amount of bytes written will be len*4

is_busy

Checks if the specified Channel is busy

wait_for

Hangs until the specified Channel becomes available. This function uses interrupts, so extra care should be taken to make sure interrupts DMA0, DMA1, DMA2 and DMA3 are enabled (irq_enable). On debug builds, this function panics if the required interrupt is not enabled, but on release this may hang for ever