Struct nds::interrupts::Flags

pub struct Flags { /* private fields */ }

Implementations

impl Flags

pub const VBLANK: Flags = Self{ bits: 1 << 0,}

pub const HBLANK: Flags = Self{ bits: 1 << 1,}

pub const VCOUNT: Flags = Self{ bits: 1 << 2,}

pub const TIMER0: Flags = Self{ bits: 1 << 3,}

pub const TIMER1: Flags = Self{ bits: 1 << 4,}

pub const TIMER2: Flags = Self{ bits: 1 << 5,}

pub const TIMER3: Flags = Self{ bits: 1 << 6,}

pub const NETWORK: Flags = Self{ bits: 1 << 7,}

pub const DMA0: Flags = Self{ bits: 1 << 8,}

pub const DMA1: Flags = Self{ bits: 1 << 9,}

pub const DMA2: Flags = Self{ bits: 1 << 10,}

pub const DMA3: Flags = Self{ bits: 1 << 11,}

pub const KEYS: Flags = Self{ bits: 1 << 12,}

pub const CART: Flags = Self{ bits: 1 << 13,}

pub const IPCSYNC: Flags = Self{ bits: 1 << 16,}

pub const FIFOEMPTY: Flags = Self{ bits: 1 << 17,}

pub const FIFONOTEMPTY: Flags = Self{ bits: 1 << 18,}

pub const CARD: Flags = Self{ bits: 1 << 19,}

pub const CARDLINE: Flags = Self{ bits: 1 << 20,}

pub const GEOMETRYFIFO: Flags = Self{ bits: 1 << 21,}

pub const LID: Flags = Self{ bits: 1 << 22,}

pub const SPI: Flags = Self{ bits: 1 << 23,}

pub const WIFI: Flags = Self{ bits: 1 << 24,}

pub const ALL: Flags = Self{ bits: (1 << 25) - 1,}

pub const fn empty() -> Flags

Returns an empty set of flags.

pub const fn all() -> Flags

Returns the set containing all flags.

pub const fn bits(&self) -> u32

Returns the raw value of the flags currently stored.

pub const fn from_bits(bits: u32) -> Option<Flags>

Convert from underlying bit representation, unless that representation contains bits that do not correspond to a flag.

pub const fn from_bits_truncate(bits: u32) -> Flags

Convert from underlying bit representation, dropping any bits that do not correspond to flags.

pub const unsafe fn from_bits_unchecked(bits: u32) -> Flags

Convert from underlying bit representation, preserving all bits (even those not corresponding to a defined flag).

Safety

The caller of the bitflags! macro can chose to allow or disallow extra bits for their bitflags type.

The caller of from_bits_unchecked() has to ensure that all bits correspond to a defined flag or that extra bits are valid for this bitflags type.

pub const fn is_empty(&self) -> bool

Returns true if no flags are currently stored.

pub const fn is_all(&self) -> bool

Returns true if all flags are currently set.

pub const fn intersects(&self, other: Flags) -> bool

Returns true if there are flags common to both self and other.

pub const fn contains(&self, other: Flags) -> bool

Returns true if all of the flags in other are contained within self.

pub fn insert(&mut self, other: Flags)

Inserts the specified flags in-place.

pub fn remove(&mut self, other: Flags)

Removes the specified flags in-place.

pub fn toggle(&mut self, other: Flags)

Toggles the specified flags in-place.

pub fn set(&mut self, other: Flags, value: bool)

Inserts or removes the specified flags depending on the passed value.

pub const fn intersection(self, other: Flags) -> Flags

Returns the intersection between the flags in self and other.

Specifically, the returned set contains only the flags which are present in both self and other.

This is equivalent to using the & operator (e.g. ops::BitAnd), as in flags & other.

pub const fn union(self, other: Flags) -> Flags

Returns the union of between the flags in self and other.

Specifically, the returned set contains all flags which are present in either self or other, including any which are present in both (see Self::symmetric_difference if that is undesirable).

This is equivalent to using the | operator (e.g. ops::BitOr), as in flags | other.

pub const fn difference(self, other: Flags) -> Flags

Returns the difference between the flags in self and other.

Specifically, the returned set contains all flags present in self, except for the ones present in other.

It is also conceptually equivalent to the โbit-clearโ operation: flags & !other (and this syntax is also supported).

This is equivalent to using the - operator (e.g. ops::Sub), as in flags - other.

pub const fn symmetric_difference(self, other: Flags) -> Flags

Returns the symmetric difference between the flags in self and other.

Specifically, the returned set contains the flags present which are present in self or other, but that are not present in both. Equivalently, it contains the flags present in exactly one of the sets self and other.

This is equivalent to using the ^ operator (e.g. ops::BitXor), as in flags ^ other.

pub const fn complement(self) -> Flags

Returns the complement of this set of flags.

Specifically, the returned set contains all the flags which are not set in self, but which are allowed for this type.

Alternatively, it can be thought of as the set difference between Self::all() and self (e.g. Self::all() - self)

This is equivalent to using the ! operator (e.g. ops::Not), as in !flags.